Create wiimotemain.cs
Note: this code uses the WiimoteLib library, which provides a C# interface for connecting to and interacting with a Wii Remote. Also note that this code only supports using the IR sensor data from the Wii Remote to control the mouse cursor, and does not include support for any other features or buttons on the Wii Remote.